+++ package.json
@@ -9,6 +9,7 @@
   },
   "license": "GPL-3.0-or-later",
   "workspaces": [
+    "packages/*",
     "plugins/*"
   ],
   "engines": {
