+++ packages/core/src/paths.ts
