+++ packages/core/src/scaffold.ts
