+++ plugins/deskwork/bin/deskwork-draft.ts
@@ -15,12 +15,12 @@
  *   { slug, stage, contentType, issueNumber?, site, calendarPath }
  */
 
-import { readConfig } from '../lib/config.ts';
-import { readCalendar, writeCalendar } from '../lib/calendar.ts';
-import { draftEntry, findEntry } from '../lib/calendar-mutations.ts';
-import { effectiveContentType } from '../lib/types.ts';
-import { resolveSite, resolveCalendarPath } from '../lib/paths.ts';
-import { absolutize, emit, fail, parseArgs } from '../lib/cli.ts';
+import { readConfig } from '@deskwork/core/config';
+import { readCalendar, writeCalendar } from '@deskwork/core/calendar';
+import { draftEntry, findEntry } from '@deskwork/core/calendar-mutations';
+import { effectiveContentType } from '@deskwork/core/types';
+import { resolveSite, resolveCalendarPath } from '@deskwork/core/paths';
+import { absolutize, emit, fail, parseArgs } from '@deskwork/core/cli-args';
 
 const KNOWN_FLAGS = ['site', 'issue'] as const;
