+++ plugins/deskwork/package.json
@@ -19,11 +19,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "rehype-stringify": "^10.0.0",
-    "remark-parse": "^11.0.0",
-    "remark-rehype": "^11.1.0",
-    "unified": "^11.0.5",
-    "yaml": "^2.8.0"
+    "@deskwork/core": "*"
   },
   "devDependencies": {
     "@types/node": "^22.10.0",
